refactor: 코드 정리
- 사용안하는 코드 제거
- 정렬 추가
