+++ src/actions/ors/ors.action.ts
@@ -2,17 +2,6 @@
 
 import {createServerSideClient} from "@/lib/supabase";
 
-// todoList 가져오기 + byId
-export const getTodoById = async (id: number) => {
-    const supabase = await createServerSideClient();
-    const result = await supabase.from("ors")
-        .select("*")
-        .is("deleted_at", null)
-        .eq("id", id);
-
-    return result.data;
-}
-
 export const createOrs = async (
     individual: number,
     interpersonal: number,
