+++ src/app/(site)/auth/error.tsx
@@ -1,19 +1,19 @@
 'use client';
 
 import React from "react";
-import {DotLoader} from "react-spinners";
+import {BounceLoader} from "react-spinners";
 
-const Loading = () => {
+const Error = () => {
     return (
         <div className="flex flex-col items-center mt-12">
             <div>
-                <DotLoader/>
+                <BounceLoader/>
             </div>
             <div className="font-bold my-2">
-                loading..
+                There is something wrong..
             </div>
         </div>
     )
 };
 
-export default Loading;
+export default Error;
